Definitions of push_lock mehtods moved to separated TU
