+++ src/mutex.cpp
@@ -1,13 +1,9 @@
-#if !defined EX_NO_PUSH_LOCKS && defined(EX_LEGACY_PUSH_LOCKS)
-#include <ntddk.h>
-#else
-#include <fltkernel.h>
-#include <ntddk.h>
-#endif
+#include <mutex.hpp>
 
 #include <ktlexcept.hpp>
-#include <mutex.hpp>
 #include <utility.hpp>
+
+#include <ntddk.h>
 
 namespace ktl {
 recursive_mutex::recursive_mutex() noexcept : MyBase() {
@@ -72,54 +68,6 @@ void shared_mutex::unlock() noexcept {
 void shared_mutex::unlock_shared() noexcept {
   ExReleaseResourceAndLeaveCriticalRegion(native_handle());
 }
-
-#ifndef KTL_EX_PUSH_LOCKS
-push_lock::push_lock() noexcept : MyBase() {
-  FltInitializePushLock(native_handle());
-}
-
-push_lock::~push_lock() noexcept {
-  FltDeletePushLock(native_handle());
-}
-
-void push_lock::lock() {
-  FltAcquirePushLockExclusive(native_handle());
-}
-
-void push_lock::lock_shared() {
-  FltAcquirePushLockShared(native_handle());
-}
-
-void push_lock::unlock() {
-  FltReleasePushLock(native_handle());
-}
-
-void push_lock::unlock_shared() {
-  FltReleasePushLock(native_handle());
-}
-#else
-push_lock::push_lock() noexcept : MyBase() {
-  ExInitializePushLock(native_handle());
-}
-
-push_lock::~push_lock() noexcept = default;
-
-void push_lock::lock() {
-  ExAcquirePushLockExclusive(native_handle());
-}
-
-void push_lock::lock_shared() {
-  ExAcquirePushLockShared(native_handle());
-}
-
-void push_lock::unlock() {
-  ExReleasePushLockExclusive(native_handle());
-}
-
-void push_lock::unlock_shared() {
-  ExReleasePushLockShared(native_handle());
-}
-#endif
 
 namespace th::details {
 void spin_lock_policy<SpinlockType::DpcOnly>::lock(
